+++ tools/settings.py
@@ -10,6 +10,31 @@ import re
 from .utils import path_from_root, exit_with_error
 from . import diagnostics
 
+LIST_SETTINGS = (
+    'TOTAL_STACK',
+    'INITIAL_MEMORY',
+    'MEMORY_GROWTH_LINEAR_STEP',
+    'MEMORY_GROWTH_GEOMETRIC_CAP',
+    'GL_MAX_TEMP_BUFFER_SIZE',
+    'MAXIMUM_MEMORY',
+    'DEFAULT_PTHREAD_STACK_SIZE'
+)
+
+COMPILE_TIME_SETTINGS = (
+    'MEMORY64',
+    'INLINING_LIMIT',
+    'DISABLE_EXCEPTION_CATCHING',
+    'EXCEPTION_CATCHING_ALLOWED',
+    'MAIN_MODULE',
+    'SIDE_MODULE',
+    'RELOCATABLE',
+    'STRICT',
+    'EMSCRIPTEN_TRACING',
+    'USE_PTHREADS',
+    'SUPPORT_LONGJMP',
+    'DEFAULT_TO_CXX',
+)
+
 
 class SettingsManager:
   attrs = {}